python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = ""
authors = [
    {name = "ben3329",email = "ben3329@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.2,<0.117.0)",
    "redis (>=6.4.0,<7.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "boto3-stubs[ssm] (>=1.40.34,<2.0.0)",
    "boto3 (>=1.40.34,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "scalar-fastapi (>=1.4.1,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "pypdf (>=6.1.0,<7.0.0)",
    "docxtpl (>=0.20.1,<0.21.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
